+++ tests/test_cbers.py
@@ -6,10 +6,10 @@ from unittest.mock import patch
 import pytest
 import rasterio
 
-from rio_tiler.errors import InvalidAssetName, MissingAssets, TileOutsideBounds
+from rio_tiler.errors import InvalidBandName, TileOutsideBounds
 from rio_tiler_pds.cbers.aws import CBERSReader
 from rio_tiler_pds.cbers.utils import sceneid_parser
-from rio_tiler_pds.errors import InvalidCBERSSceneId
+from rio_tiler_pds.errors import InvalidCBERSSceneId, MissingBands
 
 CBERS_BUCKET = os.path.join(os.path.dirname(__file__), "fixtures", "cbers-pds")
 CBERS_MUX_SCENE = "CBERS_4_MUX_20171121_057_094_L2"
@@ -33,11 +33,11 @@ def testing_env_var(monkeypatch):
     monkeypatch.setenv("GDAL_DISABLE_READDIR_ON_OPEN", "EMPTY_DIR")
 
 
-def mock_rasterio_open(asset):
+def mock_rasterio_open(band):
     """Mock rasterio Open."""
-    assert asset.startswith("s3://cbers-pds")
-    asset = asset.replace("s3://cbers-pds", CBERS_BUCKET)
-    return rasterio.open(asset)
+    assert band.startswith("s3://cbers-pds")
+    band = band.replace("s3://cbers-pds", CBERS_BUCKET)
+    return rasterio.open(band)
 
 
 @patch("rio_tiler.io.cogeo.rasterio")
@@ -56,19 +56,19 @@ def test_AWSPDS_CBERSReader_MUX(rio):
         assert len(bounds) == 4
         assert cbers.minzoom
         assert cbers.maxzoom
-        assert cbers.assets == ("B5", "B6", "B7", "B8")
+        assert cbers.bands == ("B5", "B6", "B7", "B8")
 
-        with pytest.raises(MissingAssets):
+        with pytest.raises(MissingBands):
             cbers.info()
 
-        with pytest.raises(InvalidAssetName):
-            cbers.info(assets="BAND5")
+        with pytest.raises(InvalidBandName):
+            cbers.info(bands="BAND5")
 
-        metadata = cbers.info(assets="B5")
+        metadata = cbers.info(bands="B5")
         assert len(metadata["band_metadata"]) == 1
         assert metadata["band_descriptions"] == [(1, "B5")]
 
-        metadata = cbers.info(assets=cbers.assets)
+        metadata = cbers.info(bands=cbers.bands)
         assert len(metadata["band_metadata"]) == 4
         assert metadata["band_descriptions"] == [
             (1, "B5"),
@@ -77,23 +77,23 @@ def test_AWSPDS_CBERSReader_MUX(rio):
             (4, "B8"),
         ]
 
-        with pytest.raises(MissingAssets):
+        with pytest.raises(MissingBands):
             cbers.stats()
 
-        stats = cbers.stats(assets="B5")
+        stats = cbers.stats(bands="B5")
         assert len(stats.items()) == 1
         assert stats["B5"]["pc"] == [28, 98]
 
-        stats = cbers.stats(assets=cbers.assets, hist_options=dict(bins=20))
+        stats = cbers.stats(bands=cbers.bands, hist_options=dict(bins=20))
         assert len(stats["B5"]["histogram"][0]) == 20
 
-        with pytest.raises(MissingAssets):
+        with pytest.raises(MissingBands):
             cbers.metadata()
 
-        metadata = cbers.metadata(assets="B5")
+        metadata = cbers.metadata(bands="B5")
         assert metadata["statistics"]["B5"]["pc"] == [28, 98]
 
-        metadata = cbers.metadata(assets=cbers.assets)
+        metadata = cbers.metadata(bands=cbers.bands)
         assert metadata["statistics"]["B5"]["pc"] == [28, 98]
         assert len(metadata["band_metadata"]) == 4
         assert metadata["band_descriptions"] == [
@@ -106,9 +106,7 @@ def test_AWSPDS_CBERSReader_MUX(rio):
         tile_z = 10
         tile_x = 664
         tile_y = 495
-        data, mask = cbers.tile(
-            tile_x, tile_y, tile_z, assets=cbers.scene_params["rgb"]
-        )
+        data, mask = cbers.tile(tile_x, tile_y, tile_z, bands=cbers.scene_params["rgb"])
         assert data.shape == (3, 256, 256)
         assert mask.shape == (256, 256)
 
@@ -116,7 +114,7 @@ def test_AWSPDS_CBERSReader_MUX(rio):
         tile_x = 694
         tile_y = 495
         with pytest.raises(TileOutsideBounds):
-            cbers.tile(tile_x, tile_y, tile_z, assets=cbers.scene_params["rgb"])
+            cbers.tile(tile_x, tile_y, tile_z, bands=cbers.scene_params["rgb"])
 
         tile_z = 10
         tile_x = 664
@@ -139,14 +137,12 @@ def test_AWSPDS_CBERSReader_AWFI(rio):
         assert len(bounds) == 4
         assert cbers.minzoom
         assert cbers.maxzoom
-        assert cbers.assets == ("B13", "B14", "B15", "B16")
+        assert cbers.bands == ("B13", "B14", "B15", "B16")
 
         tile_z = 10
         tile_x = 401
         tile_y = 585
-        data, mask = cbers.tile(
-            tile_x, tile_y, tile_z, assets=cbers.scene_params["rgb"]
-        )
+        data, mask = cbers.tile(tile_x, tile_y, tile_z, bands=cbers.scene_params["rgb"])
         assert data.shape == (3, 256, 256)
         assert mask.shape == (256, 256)
 
@@ -162,14 +158,12 @@ def test_AWSPDS_CBERSReader_PAN10M(rio):
         assert len(bounds) == 4
         assert cbers.minzoom
         assert cbers.maxzoom
-        assert cbers.assets == ("B2", "B3", "B4")
+        assert cbers.bands == ("B2", "B3", "B4")
 
         tile_z = 10
         tile_x = 370
         tile_y = 535
-        data, mask = cbers.tile(
-            tile_x, tile_y, tile_z, assets=cbers.scene_params["rgb"]
-        )
+        data, mask = cbers.tile(tile_x, tile_y, tile_z, bands=cbers.scene_params["rgb"])
         assert data.shape == (3, 256, 256)
         assert mask.shape == (256, 256)
 
@@ -185,14 +179,12 @@ def test_AWSPDS_CBERSReader_PAN5M(rio):
         assert len(bounds) == 4
         assert cbers.minzoom
         assert cbers.maxzoom
-        assert cbers.assets == ("B1",)
+        assert cbers.bands == ("B1",)
 
         tile_z = 10
         tile_x = 390
         tile_y = 547
-        data, mask = cbers.tile(
-            tile_x, tile_y, tile_z, assets=cbers.scene_params["rgb"]
-        )
+        data, mask = cbers.tile(tile_x, tile_y, tile_z, bands=cbers.scene_params["rgb"])
         assert data.shape == (3, 256, 256)
         assert mask.shape == (256, 256)