+++ tests/test_landsat.py
@@ -7,8 +7,8 @@ import numpy
 import pytest
 import rasterio
 
-from rio_tiler.errors import InvalidAssetName, MissingAssets, TileOutsideBounds
-from rio_tiler_pds.errors import InvalidLandsatSceneId
+from rio_tiler.errors import InvalidBandName, TileOutsideBounds
+from rio_tiler_pds.errors import InvalidLandsatSceneId, MissingBands
 from rio_tiler_pds.landsat.aws import L8Reader
 from rio_tiler_pds.landsat.utils import sceneid_parser
 
@@ -33,11 +33,11 @@ def testing_env_var(monkeypatch):
     monkeypatch.setenv("GDAL_DISABLE_READDIR_ON_OPEN", "TRUE")
 
 
-def mock_rasterio_open(asset):
+def mock_rasterio_open(band):
     """Mock rasterio Open."""
-    assert asset.startswith("s3://landsat-pds")
-    asset = asset.replace("s3://landsat-pds", LANDSAT_BUCKET)
-    return rasterio.open(asset)
+    assert band.startswith("s3://landsat-pds")
+    band = band.replace("s3://landsat-pds", LANDSAT_BUCKET)
+    return rasterio.open(band)
 
 
 @patch("rio_tiler_pds.reader.aws_get_object")
@@ -56,7 +56,7 @@ def test_AWSPDS_L8Reader(rio, get_object):
         assert landsat.minzoom == 7
         assert landsat.maxzoom == 12
         assert len(landsat.bounds) == 4
-        assert landsat.assets == (
+        assert landsat.bands == (
             "B1",
             "B2",
             "B3",
@@ -71,82 +71,80 @@ def test_AWSPDS_L8Reader(rio, get_object):
             "BQA",
         )
 
-        with pytest.raises(MissingAssets):
+        with pytest.raises(MissingBands):
             landsat.info()
 
-        with pytest.raises(InvalidAssetName):
-            landsat.info(assets="BAND5")
+        with pytest.raises(InvalidBandName):
+            landsat.info(bands="BAND5")
 
-        metadata = landsat.info(assets="B5")
+        metadata = landsat.info(bands="B5")
         assert len(metadata["band_metadata"]) == 1
         assert metadata["band_descriptions"] == [(1, "B5")]
 
-        metadata = landsat.info(assets=landsat.assets)
+        metadata = landsat.info(bands=landsat.bands)
         assert len(metadata["band_metadata"]) == 12
 
-        with pytest.raises(MissingAssets):
+        with pytest.raises(MissingBands):
             landsat.stats()
 
-        stats = landsat.stats(assets="B1")
+        stats = landsat.stats(bands="B1")
         assert stats["B1"]["pc"] == [1206, 6957]
 
-        stats = landsat.stats(assets=landsat.assets)
+        stats = landsat.stats(bands=landsat.bands)
         assert len(stats.items()) == 12
-        assert list(stats) == list(landsat.assets)
+        assert list(stats) == list(landsat.bands)
 
-        stats = landsat.stats(assets="B1", hist_options=dict(bins=20))
+        stats = landsat.stats(bands="B1", hist_options=dict(bins=20))
         assert len(stats["B1"]["histogram"][0]) == 20
 
-        stats = landsat.stats(pmin=10, pmax=90, assets="B1")
+        stats = landsat.stats(pmin=10, pmax=90, bands="B1")
         assert stats["B1"]["pc"] == [1274, 3964]
 
-        stats = landsat.stats(assets="BQA", nodata=0, resampling_method="bilinear")
+        stats = landsat.stats(bands="BQA", nodata=0, resampling_method="bilinear")
         assert stats["BQA"]["min"] == 2720
 
-        with pytest.raises(MissingAssets):
+        with pytest.raises(MissingBands):
             landsat.metadata()
 
-        metadata = landsat.metadata(assets="B1")
+        metadata = landsat.metadata(bands="B1")
         assert metadata["statistics"]["B1"]["pc"] == [1206, 6957]
         assert metadata["band_metadata"] == [(1, {})]
         assert metadata["band_descriptions"] == [(1, "B1")]
 
-        metadata = landsat.metadata(assets=("B1", "B2"))
+        metadata = landsat.metadata(bands=("B1", "B2"))
         assert metadata["band_metadata"] == [(1, {}), (2, {})]
         assert metadata["band_descriptions"] == [(1, "B1"), (2, "B2")]
 
-        metadata = landsat.metadata(
-            assets="BQA", nodata=0, resampling_method="bilinear"
-        )
+        metadata = landsat.metadata(bands="BQA", nodata=0, resampling_method="bilinear")
         assert metadata["statistics"]["BQA"]["min"] == 2720
 
         tile_z = 8
         tile_x = 71
         tile_y = 102
 
-        with pytest.raises(MissingAssets):
+        with pytest.raises(MissingBands):
             landsat.tile(tile_x, tile_y, tile_z)
 
-        data, mask = landsat.tile(tile_x, tile_y, tile_z, assets=("B4", "B3", "B2"))
+        data, mask = landsat.tile(tile_x, tile_y, tile_z, bands=("B4", "B3", "B2"))
         assert data.shape == (3, 256, 256)
         assert data.dtype == numpy.uint16
         assert mask.shape == (256, 256)
         assert not mask.all()
 
         # Temp are float32
-        data, mask = landsat.tile(tile_x, tile_y, tile_z, assets="B10")
+        data, mask = landsat.tile(tile_x, tile_y, tile_z, bands="B10")
         assert data.shape == (1, 256, 256)
         assert data.dtype == numpy.float32
         assert mask.shape == (256, 256)
 
         data, mask = landsat.tile(
-            tile_x, tile_y, tile_z, assets="BQA", nodata=0, resampling_method="bilinear"
+            tile_x, tile_y, tile_z, bands="BQA", nodata=0, resampling_method="bilinear"
         )
         assert data.shape == (1, 256, 256)
         assert not mask.all()
 
         data, mask = landsat.tile(
-            tile_x, tile_y, tile_z, assets=("B4", "B3", "B2"), pan=True
+            tile_x, tile_y, tile_z, bands=("B4", "B3", "B2"), pan=True
         )
         assert data.shape == (3, 256, 256)
         assert data.dtype == numpy.uint16
@@ -156,7 +154,7 @@ def test_AWSPDS_L8Reader(rio, get_object):
         tile_x = 701
         tile_y = 102
         with pytest.raises(TileOutsideBounds):
-            landsat.tile(tile_x, tile_y, tile_z, assets=("B4", "B3", "B2"))
+            landsat.tile(tile_x, tile_y, tile_z, bands=("B4", "B3", "B2"))
 
         tile_z = 8
         tile_x = 71
@@ -168,23 +166,23 @@ def test_AWSPDS_L8Reader(rio, get_object):
         assert data.dtype == numpy.float64
         assert mask.shape == (256, 256)
 
-        with pytest.raises(MissingAssets):
+        with pytest.raises(MissingBands):
             landsat.preview()
 
-        data, mask = landsat.preview(assets=("B4", "B3", "B2"))
+        data, mask = landsat.preview(bands=("B4", "B3", "B2"))
         assert data.shape == (3, 259, 255)
         assert data.dtype == numpy.uint16
         assert mask.shape == (259, 255)
         assert not mask.all()
 
         # Temp are float32
-        data, mask = landsat.preview(assets="B10")
+        data, mask = landsat.preview(bands="B10")
         assert data.shape == (1, 259, 255)
         assert data.dtype == numpy.float32
         assert mask.shape == (259, 255)
 
         data, mask = landsat.preview(
-            assets=("B4", "B3", "B2"), pan=True, width=256, height=256
+            bands=("B4", "B3", "B2"), pan=True, width=256, height=256
         )
         assert data.shape == (3, 256, 256)
         assert data.dtype == numpy.uint16
@@ -196,30 +194,30 @@ def test_AWSPDS_L8Reader(rio, get_object):
         assert mask.shape == (259, 255)
 
         data, mask = landsat.preview(
-            assets="BQA", nodata=0, resampling_method="bilinear"
+            bands="BQA", nodata=0, resampling_method="bilinear"
         )
         assert data.shape == (1, 259, 255)
         assert not mask.all()
 
-        with pytest.raises(MissingAssets):
+        with pytest.raises(MissingBands):
             landsat.point(-80.094, 33.2062)
 
-        values = landsat.point(-80.094, 33.2062, assets="B7")
+        values = landsat.point(-80.094, 33.2062, bands="B7")
         assert values == [667]
 
-        values = landsat.point(-80.094, 33.2062, assets="BQA")
+        values = landsat.point(-80.094, 33.2062, bands="BQA")
         assert values[0] == 2800
 
-        values = landsat.point(-80.094, 33.2062, assets=("B7", "B4"))
+        values = landsat.point(-80.094, 33.2062, bands=("B7", "B4"))
         assert len(values) == 2
 
         values = landsat.point(-80.094, 33.2062, expression="B5*0.8, B4*1.1, B3*0.8")
         assert len(values) == 3
 
-        with pytest.raises(MissingAssets):
+        with pytest.raises(MissingBands):
             landsat.part((-80.593, 32.9134, -79.674, 33.6790))
 
-        data, mask = landsat.part((-80.593, 32.9134, -79.674, 33.6790), assets="B7")
+        data, mask = landsat.part((-80.593, 32.9134, -79.674, 33.6790), bands="B7")
         assert data.shape == (1, 87, 104)
         assert data.dtype == numpy.uint16
         assert mask.shape == (87, 104)
@@ -227,7 +225,7 @@ def test_AWSPDS_L8Reader(rio, get_object):
 
         data, _ = landsat.part(
             (-80.593, 32.9134, -79.674, 33.6790),
-            assets="BQA",
+            bands="BQA",
             nodata=0,
             resampling_method="bilinear",
         )
@@ -243,7 +241,7 @@ def test_AWSPDS_L8Reader(rio, get_object):
 
         data, mask = landsat.part(
             (-80.593, 32.9134, -79.674, 33.6790),
-            assets=("B4", "B3", "B2"),
+            bands=("B4", "B3", "B2"),
             pan=True,
             width=80,
             height=80,